Был удалён класс Creator, классы Database и FamilyTree в состоянии сами создавать новые классы. Добавлен базовый интерфейс UIView, от которого наследуются интерфейсы DatabaseView и FamilyTreeView. На интерфейс UIView подписан базовый класс UI.
