+++ src/Model/Creatures/Human.java
@@ -17,12 +17,6 @@ public class Human implements Serializable, Creature<Human> {
     private final ArrayList<Human> children = new ArrayList<>();
     private Human mather = null, father = null, spouse = null;
 
-    public Human(String name, String surname, Gender gender){
-        this.name = name;
-        this.surname = surname;
-        this.gender = gender;
-    }
-
     public String getName() {
         return name;
     }
