+++ src/Model/Database.java
@@ -1,28 +1,26 @@
 package Model;
 
+import Model.Creatures.Human;
+import Model.FamilyTree.FamilyTree;
 import Model.Interface.DatabaseInterface;
 import Model.Interface.FamilyTreeInterface;
 
 import java.io.Serializable;
-import java.util.ArrayList;
 
 public class Database implements DatabaseInterface, Serializable {
-    public static final String SUCCESSFUL = "Successful";
-    public static final String OBJECT_EXIST = "This object already exists";
-
-    private final Creator creator = new Creator();
-
+    public enum creatureType {
+        human
+    }
 
     public FamilyTreeInterface<?> createNewFamilyTree(String nameCreature) {
-        return this.creator.createNewFamilyTree(nameCreature);
+        if (nameCreature.equalsIgnoreCase(creatureType.human.toString()))
+            return new FamilyTree<Human>(Human::new);
+
+        return null;
     }
 
     public FamilyTreeInterface<?> getFamilyTree(String name){
         return null;
-    }
-
-    public ArrayList<String> getAvailableCreatures() {
-        return this.creator.getAvailableCreatures();
     }
 }
