+++ src/View/Interface/DatabaseView.java
@@ -2,12 +2,6 @@ package View.Interface;
 
 import Presenter.DatabasePresenter.DatabasePresenter;
 
-public interface DatabaseView {
+public interface DatabaseView extends UIView{
     void setPresenter(DatabasePresenter databasePresenter);
-    void mainLoop();
-    void exit();
-    String nextLine();
-    void println(String string);
-    void printMessage(String message);
-    void printSignedMessage(String name, String message);
 }
