+++ src/View/Interface/FamilyTreeView.java
@@ -2,12 +2,6 @@ package View.Interface;
 
 import Presenter.FamilyTreePresenter.FamilyTreePresenter;
 
-public interface FamilyTreeView {
+public interface FamilyTreeView extends UIView {
     void setPresenter(FamilyTreePresenter presenter);
-    void mainLoop();
-    void exit();
-    String nextLine();
-    void println(String string);
-    void printMessage(String message);
-    void printSignedMessage(String name, String message);
 }
